PR feedback + Moved Image module to separate file
