+++ vnext/ReactUWP/ReactUWP.vcxproj.filters
@@ -228,6 +228,9 @@
     <ClCompile Include="Views\ExpressionAnimationStore.cpp">
       <Filter>Views</Filter>
     </ClCompile>
+    <ClCompile Include="Modules\ImageViewManagerModule.cpp">
+      <Filter>Modules</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="Modules\DeviceInfoModule.h">
@@ -462,9 +465,6 @@
     <ClInclude Include="Views\cppwinrt\Microsoft.UI.Composition.Effects.BorderEffect.g.h">
       <Filter>Views\cppwinrt</Filter>
     </ClInclude>
-    <ClInclude Include="Views\Image\BorderEffect.h">
-      <Filter>Views\Image</Filter>
-    </ClInclude>
     <ClInclude Include="Views\Image\ReactImageBrush.h">
       <Filter>Views\Image</Filter>
     </ClInclude>
@@ -473,6 +473,15 @@
     </ClInclude>
     <ClInclude Include="Views\ExpressionAnimationStore.h">
       <Filter>Views</Filter>
+    </ClInclude>
+    <ClInclude Include="Modules\ImageViewManagerModule.h">
+      <Filter>Modules</Filter>
+    </ClInclude>
+    <ClInclude Include="Views\Image\Microsoft.UI.Composition.Effects_Impl.h">
+      <Filter>Views\Image</Filter>
+    </ClInclude>
+    <ClInclude Include="Views\Image\BorderEffect.h">
+      <Filter>Views\Image</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
