+++ vnext/ReactUWP/Views/Image/ImageViewManager.cpp
@@ -2,35 +2,22 @@
 // Licensed under the MIT License.
 
 // NYI:
-//   implement image cache
-//   implement prefetch functionality
 //   implement multi source (parse out most suitable image source from array of sources)
-
 #include "pch.h"
 
 #include "ImageViewManager.h"
 
 #include <winrt/Windows.Foundation.h>
 #include <winrt/Windows.UI.Xaml.Controls.h>
-#include <winrt/Windows.UI.Xaml.Media.Imaging.h>
-
-#include <cxxreact/JsArgumentHelpers.h>
 
 #include <IReactInstance.h>
 #include <Views/ShadowNodeBase.h>
 #include <Utils/PropertyHandlerUtils.h>
 #include "ReactImage.h"
 
-#if _MSC_VER <= 1913
-// VC 19 (2015-2017.6) cannot optimize co_await/cppwinrt usage
-#pragma optimize( "", off )
-#endif
-
 namespace winrt {
   using namespace Windows::Foundation;
-  using namespace Windows::Storage::Streams;
   using namespace Windows::UI::Xaml::Controls;
-  using namespace Windows::UI::Xaml::Media::Imaging;
 }
 
 // Such code is better to move to a seperate parser layer
@@ -66,7 +53,7 @@ struct json_type_traits<react::uwp::ResizeMode>
 {
   static react::uwp::ResizeMode parseJson(const folly::dynamic& json)
   {
-    auto resizeMode = react::uwp::ResizeMode::Contain;
+    auto resizeMode{ react::uwp::ResizeMode::Contain };
 
     if (json == "cover")
     {
@@ -109,7 +96,7 @@ namespace react { namespace uwp {
   {
     auto reactImage{ ReactImage::Create() };
 
-    reactImage->OnLoadEnd([=](const auto&, const bool& succeeded)
+    reactImage->OnLoadEnd([this, reactImage](const auto&, const bool& succeeded)
     {
         ImageSource source{ reactImage->Source() };
 
@@ -122,15 +109,15 @@ namespace react { namespace uwp {
 
   void ImageViewManager::UpdateProperties(ShadowNodeBase* nodeToUpdate, const folly::dynamic& reactDiffMap)
   {
-    auto canvas = nodeToUpdate->GetView().as<winrt::Canvas>();
+    auto canvas{ nodeToUpdate->GetView().as<winrt::Canvas>() };
 
     if (canvas == nullptr)
       return;
 
     for (const auto& pair : reactDiffMap.items())
     {
-      const std::string& propertyName = pair.first.getString();
-      const folly::dynamic& propertyValue = pair.second;
+      const std::string& propertyName{ pair.first.getString() };
+      const folly::dynamic& propertyValue{ pair.second };
 
       if (propertyName == "source")
       {
@@ -138,7 +125,7 @@ namespace react { namespace uwp {
       }
       else if (propertyName == "resizeMode")
       {
-        auto resizeMode = json_type_traits<react::uwp::ResizeMode>::parseJson(propertyValue);
+        auto resizeMode{ json_type_traits<react::uwp::ResizeMode>::parseJson(propertyValue) };
         auto reactImage{ canvas.as<ReactImage>() };
         reactImage->ResizeMode(resizeMode);
       }
@@ -168,20 +155,13 @@ namespace react { namespace uwp {
 
   void ImageViewManager::setSource(winrt::Canvas canvas, const folly::dynamic& data)
   {
-    auto instance = m_wkReactInstance.lock();
+    auto instance{ m_wkReactInstance.lock() };
     if (instance == nullptr)
       return;
 
-    auto sources = json_type_traits<std::vector<ImageSource>>::parseJson(data);
-
-    auto uriString = sources[0].uri;
-    if (uriString.length() == 0)
-    {
-      EmitImageEvent(instance, canvas, "topError", sources[0]);
-      return;
-    }
-
+    auto sources{ json_type_traits<std::vector<ImageSource>>::parseJson(data) };
     auto reactImage{ canvas.as<ReactImage>() };
+
     EmitImageEvent(instance, canvas, "topLoadStart", sources[0]);
     reactImage->Source(sources[0]);
   }
@@ -208,127 +188,4 @@ namespace react { namespace uwp {
 
     return props;
   }
-
-
-//
-// ImageViewManagerModule::ImageViewManagerModuleImpl
-//
-class ImageViewManagerModule::ImageViewManagerModuleImpl
-{
-public:
-  ImageViewManagerModuleImpl(ImageViewManagerModule* parent, const std::shared_ptr<facebook::react::MessageQueueThread>& defaultQueueThread)
-    : m_parent(parent)
-    , m_queueThread(defaultQueueThread)
-  { }
-
-  void Disconnect()
-  {
-    m_parent = nullptr;
-  }
-
-  void getSize(std::string uri, Callback successCallback, Callback errorCallback);
-  void prefetchImage(std::string uri, Callback successCallback, Callback errorCallback);
-  void queryCache(const folly::dynamic& requests, Callback successCallback, Callback errorCallback);
-
-private:
-  ImageViewManagerModule *m_parent;
-  std::shared_ptr<facebook::react::MessageQueueThread> m_queueThread;
-};
-
-winrt::fire_and_forget GetImageSizeAsync(std::string uri, facebook::xplat::module::CxxModule::Callback successCallback, facebook::xplat::module::CxxModule::Callback errorCallback)
-{
-  bool succeeded{ false };
-
-  try
-  {
-    ImageSource source;
-    source.uri = uri;
-
-    winrt::InMemoryRandomAccessStream memoryStream{ co_await react::uwp::GetImageStreamAsync(source) };
-
-    if (!memoryStream)
-    {
-      EmitImageEvent(std::weak_ptr<IReactInstance>().lock(), {}, "topError", source);
-    }
-
-    winrt::BitmapImage bitmap;
-    co_await bitmap.SetSourceAsync(memoryStream);
-
-    if (bitmap)
-    {
-      successCallback({ bitmap.PixelWidth(), bitmap.PixelHeight() });
-      succeeded = true;
-    }
-  }
-  catch (winrt::hresult_error const &)
-  {
-  }
-
-  if (!succeeded)
-    errorCallback({});
-
-  co_return;
-}
-
-void ImageViewManagerModule::ImageViewManagerModuleImpl::getSize(std::string uri, Callback successCallback, Callback errorCallback)
-{
-  GetImageSizeAsync(uri, successCallback, errorCallback);
-}
-
-void ImageViewManagerModule::ImageViewManagerModuleImpl::prefetchImage(std::string /*uri*/, Callback successCallback, Callback /*errorCallback*/)
-{
-  // NotYetImplemented
-  successCallback({});
-}
-
-void ImageViewManagerModule::ImageViewManagerModuleImpl::queryCache(const folly::dynamic& requests, Callback successCallback, Callback /*errorCallback*/)
-{
-  // NotYetImplemented
-  successCallback({ folly::dynamic::object() });
-}
-
-
-//
-// ImageViewManagerModule
-//
-const char* ImageViewManagerModule::name = "ImageViewManager";
-
-ImageViewManagerModule::ImageViewManagerModule(const std::shared_ptr<facebook::react::MessageQueueThread>& defaultQueueThread)
-  : m_imageViewManagerModule(std::make_shared<ImageViewManagerModuleImpl>(this, defaultQueueThread))
-{
-}
-
-ImageViewManagerModule::~ImageViewManagerModule()
-{
-}
-
-std::string ImageViewManagerModule::getName()
-{
-  return name;
-}
-
-std::map<std::string, folly::dynamic> ImageViewManagerModule::getConstants()
-{
-  return { {} };
-}
-
-auto ImageViewManagerModule::getMethods() -> std::vector<Method>
-{
-  std::shared_ptr<ImageViewManagerModuleImpl> imageViewManager(m_imageViewManagerModule);
-  return {
-    Method("getSize", [imageViewManager](folly::dynamic args, Callback successCallback, Callback errorCallback)
-    {
-      imageViewManager->getSize(facebook::xplat::jsArgAsString(args, 0), successCallback, errorCallback);
-    }, AsyncTag),
-    Method("prefetchImage", [imageViewManager](folly::dynamic args, Callback successCallback, Callback errorCallback)
-    {
-      imageViewManager->prefetchImage(facebook::xplat::jsArgAsString(args, 0), successCallback, errorCallback);
-    }),
-    Method("queryCache", [imageViewManager](folly::dynamic args, Callback successCallback, Callback errorCallback)
-    {
-      imageViewManager->queryCache(args[0], successCallback, errorCallback);
-    }),
-  };
-}
-
 } }
