+++ vnext/ReactUWP/Views/Image/ImageViewManager.h
@@ -1,13 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-#include <cxxreact/CxxModule.h>
 #include <Views/FrameworkElementViewManager.h>
-#include <winrt/Windows.UI.Xaml.Media.h>
-
-namespace facebook { namespace react {
-  class MessageQueueThread;
-} }
 
 namespace react { namespace uwp {
 
@@ -29,23 +23,4 @@ namespace react { namespace uwp {
   private:
     void setSource(winrt::Windows::UI::Xaml::Controls::Canvas canvas, const folly::dynamic& sources);
   };
-
-  class ImageViewManagerModule : public facebook::xplat::module::CxxModule
-  {
-  public:
-    ImageViewManagerModule(const std::shared_ptr<facebook::react::MessageQueueThread>& defaultQueueThread);
-    virtual ~ImageViewManagerModule();
-
-    // CxxModule
-    std::string getName() override;
-    std::map<std::string, folly::dynamic> getConstants() override;
-    auto getMethods() -> std::vector<Method> override;
-
-    static const char* name;
-
-  private:
-    class ImageViewManagerModuleImpl;
-    std::shared_ptr<ImageViewManagerModuleImpl> m_imageViewManagerModule;
-  };
-
 } }
